requirements.txt
google-generativeai
httpx
pypdf
tiktoken
Flask
Markdown
python-dotenv
requests
beautifulsoup4
selenium
gunicorn 